c rigolo ca fonctionne presque

+++ Assets/Scriptes/Bubble.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 
 public class Bubble : MonoBehaviour {
-    public float prix;
+    public int prix;
     public Texture2D sprite;
 }
